refactor(suite-common): share wrapped native pending tx tracking
